wip: Provide basic scripts to create/backup/restore Postgis/Mariadb databases #230 [skip ci]

+++ scripts/database.sh
@@ -1,21 +1,46 @@
 #!/bin/bash
 set -euo pipefail
 
+exist_postgis_db() {
+  local DATABASE="$1"
+  local DOCKER_RUN="docker run --rm --network=${DOCKER_BACK_NETWORK} ${POSTGIS_IMAGE}:${POSTGIS_TAG}"
+  local PSQL="psql -d postgresql://${POSTGIS_USER}:${POSTGIS_PASSWORD}@postgis"
+  
+  local DATABASE_EXISTS=`${DOCKER_RUN} ${PSQL} -tc "SELECT 1 FROM pg_database WHERE datname = '${DATABASE}'"`
+  if [ -z "${DATABASE_EXISTS}" ]; then
+    return 0
+  else
+    return 1
+  fi
+}
+
+exist_mariadb_db() {
+  local DATABASE=$1
+  local DOCKER_RUN="docker run --rm --network=${DOCKER_BACK_NETWORK} ${MARIADB_IMAGE}:${MARIADB_TAG}"
+  local MYSQL="mysql --host=mariadb --password=${MARIADB_ROOT_PASSWORD}"
+
+  local DATABASE_EXISTS=`${DOCKER_RUN} ${MYSQL} -e "SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME='${DATABASE}'"`
+  if [ -z "${DATABASE_EXISTS}" ]; then
+    return 0
+  else
+    return 1
+  fi
+}
+
 create_postgis_db() {
-  local USER=$1
-  local PASSWORD=$2
-  local DATABASE=$3
+  local USER="$1"
+  local PASSWORD="$2"
+  local DATABASE="$3"
   local DOCKER_RUN="docker run --rm --network=${DOCKER_BACK_NETWORK} ${POSTGIS_IMAGE}:${POSTGIS_TAG}"
   local PSQL="psql -d postgresql://${POSTGIS_USER}:${POSTGIS_PASSWORD}@postgis"
 
-  local DATABASE_EXISTS=`${DOCKER_RUN} ${PSQL} -tc "SELECT 1 FROM pg_database WHERE datname = '${DATABASE}'"`
-  if [ -z "${DATABASE_EXISTS}" ]; then
+  if ! exist_postgis_db ${DATABASE}; then
     echo creating database \"${DATABASE}\"
     ${DOCKER_RUN} ${PSQL} -c "CREATE DATABASE ${DATABASE};"
     ${DOCKER_RUN} ${PSQL} -c "CREATE USER ${USER} WITH ENCRYPTED PASSWORD '${PASSWORD}';"
     ${DOCKER_RUN} ${PSQL} -c "GRANT ALL PRIVILEGES ON DATABASE ${DATABASE} TO ${USER};"
   else
-    echo the database \"${DATABASE}\" already exists
+    echo warning: the specified database \"${DATABASE}\" already exist
   fi
 }
 
@@ -26,14 +51,13 @@ create_mariadb_db() {
   local DOCKER_RUN="docker run --rm --network=${DOCKER_BACK_NETWORK} ${MARIADB_IMAGE}:${MARIADB_TAG}"
   local MYSQL="mysql --host=mariadb --password=${MARIADB_ROOT_PASSWORD}"
 
-  local DATABASE_EXISTS=`${DOCKER_RUN} ${MYSQL} -e "SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME='${DATABASE}'"`
-  if [ -z "${DATABASE_EXISTS}" ]; then
+  if ! exist_mariadb_db ${DATABASE}; then
     echo creating database \"${DATABASE}\"
     ${DOCKER_RUN} ${MYSQL} -e "CREATE DATABASE ${DATABASE};"
     ${DOCKER_RUN} ${MYSQL} -e "GRANT ALL PRIVILEGES ON ${DATABASE}.* TO '${USER}'@'%' IDENTIFIED BY '${PASSWORD}';"
     ${DOCKER_RUN} ${MYSQL} -e "FLUSH PRIVILEGES;"
   else
-    echo the database \"${DATABASE}\" already exists
+    echo warning: the specified database \"${DATABASE}\" already exist
   fi
 }
 
@@ -43,8 +67,13 @@ drop_postgis_db() {
   local DOCKER_RUN="docker run --rm --network=${DOCKER_BACK_NETWORK} ${POSTGIS_IMAGE}:${POSTGIS_TAG}"
   local PSQL="psql -d postgresql://${POSTGIS_USER}:${POSTGIS_PASSWORD}@postgis"
 
-  ${DOCKER_RUN} ${PSQL} -c "DROP DATABASE ${DATABASE}"
-  ${DOCKER_RUN} ${PSQL} -c "DROP ROLE ${USER}"
+  if exist_postgis_db ${DATABASE}; then
+    echo droping database \"${DATABASE}\"
+    ${DOCKER_RUN} ${PSQL} -c "DROP DATABASE ${DATABASE}"
+    ${DOCKER_RUN} ${PSQL} -c "DROP ROLE ${USER}"
+  else
+    echo waarning: the specified database \"${DATABASE}\" doest not exist
+  fi
 }
 
 drop_mariadb_db() {
@@ -53,8 +82,13 @@ drop_mariadb_db() {
   local DOCKER_RUN="docker run --rm --network=${DOCKER_BACK_NETWORK} ${MARIADB_IMAGE}:${MARIADB_TAG}"
   local MYSQL="mysql --host=mariadb --password=${MARIADB_ROOT_PASSWORD}"
 
-  ${DOCKER_RUN} ${MYSQL} -e "DROP DATABASE ${DATABASE};"
-  ${DOCKER_RUN} ${MYSQL} -e "DROP USER ${USER}"
+  if exists_mariadb_db ${DATABASE}; then
+    echo droping database \"${DATABASE}\"
+    ${DOCKER_RUN} ${MYSQL} -e "DROP DATABASE ${DATABASE};"
+    ${DOCKER_RUN} ${MYSQL} -e "DROP USER ${USER}"
+  else
+    echo warning: the specified database \"${DATABASE}\" doest not exist    
+  fi
 }
 
 backup_postgis_db() {
